renamed files + updated Postman collection

+++ internal/nbi/nbi-utils.go
@@ -13,7 +13,58 @@ type GatewayInfo struct {
 	RouterID  string `json:"routerID" binding:"required"`
 }
 
-// clean up procedure
+// GET /management/db
+func (env *Env) GetDB(c *gin.Context) {
+	log.Info("GetDB called: retrieving all GatewayConnectivity instances")
+	c.JSON(http.StatusOK, gin.H{"gateways": env.DB})
+}
+
+// POST /management/db
+func (env *Env) AddDBEntry(c *gin.Context) {
+	log.Info("AddDBEntry: called")
+	var json GatewayConnectivity
+	if err := c.ShouldBindJSON(&json); err != nil {
+		log.Error("JSON body not well formatted")
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	sliceId := json.SliceID
+	networkId := json.PrivNetID
+	subnetId := json.SubnetID
+	routerId := json.RouterID
+	log.Info("Adding new GatewayConnectivity istance with infos ", json)
+	log.Info(sliceId, networkId, subnetId, routerId)
+
+	index, _, err := env.RetrieveGatewayConnectivityFromDB(sliceId)
+	if err != nil && index == -1 {
+		env.DB = append(env.DB, json)
+		log.Info("Added a new GatewayConnectivty for slice with sliceID: ", sliceId, " \n[DB len: ", len(env.DB), " capacity: ", cap(env.DB), "] \nDB:", env.DB)
+		c.JSON(http.StatusOK, gin.H{"gateway": json})
+		return
+	}
+	c.Status(http.StatusBadRequest)
+}
+
+// DELETE /management/db?sliceId=test
+func (env *Env) DeleteDBEntry(c *gin.Context) {
+	sliceId := c.Query("sliceId")
+	log.Info("DeleteDBEntry: called with param sliceID: " + sliceId)
+	if sliceId != "" {
+		_, _, err := env.RetrieveGatewayConnectivityFromDB(sliceId)
+		if err != nil {
+			c.Status(http.StatusNotFound)
+		}
+		env.RemoveGatewayConnectivityFromDB(sliceId)
+		c.Status(http.StatusOK)
+	}
+	c.Status(http.StatusBadRequest)
+}
+
+//
+// Managementc clean functions
+//
+
+// GET /management/clean clean up procedure
 func (env *Env) CleanUp(c *gin.Context) {
 	for i := range env.DB {
 		obj := env.DB[i]
@@ -27,7 +78,8 @@ func (env *Env) CleanUp(c *gin.Context) {
 	c.Status(http.StatusOK)
 }
 
-// clean up procedure
+// POST /management/clean clean up procedure with
+// gatewayInfo as body. It deletes the resources passed as params
 func (env *Env) CleanUpGateway(c *gin.Context) {
 	var json GatewayInfo
 	if err := c.ShouldBindJSON(&json); err != nil {
@@ -41,7 +93,7 @@ func (env *Env) CleanUpGateway(c *gin.Context) {
 	routerId := json.RouterID
 	log.Info("Clean procedure...")
 
-	if routerId != "" {
+	if routerId != "" && subnetId != "" {
 		log.Info("Removing router with ID ", routerId)
 		env.Client.DeleteRouter(routerId, subnetId)
 	}
